+++ tests/test_on_data/test_recording_interfaces.py
@@ -6,6 +6,7 @@ from unittest import skip, skipIf
 import jsonschema
 import numpy as np
 from hdmf.testing import TestCase
+from numpy.testing import assert_array_equal
 from packaging import version
 from pynwb import NWBHDF5IO
 
@@ -211,7 +212,8 @@ class TestMEArecRecordingInterface(RecordingExtractorInterfaceTestMixin, TestCas
         assert len(metadata["Ecephys"]["Device"]) == 1
         assert metadata["Ecephys"]["Device"][0]["name"] == "Neuronexus-32"
         assert metadata["Ecephys"]["Device"][0]["description"] == "The ecephys device for the MEArec recording."
-        # assert len(metadata["Ecephys"]["ElectrodeGroup"]) == 1 # do not test this condition because in the test we are setting a mock probe
+        # assert len(metadata["Ecephys"]["ElectrodeGroup"]) == 1
+        # do not test this condition because in the test we are setting a mock probe
         assert metadata["Ecephys"]["ElectrodeGroup"][0]["device"] == "Neuronexus-32"
         assert metadata["Ecephys"]["ElectricalSeries"]["description"] == (
             '{"angle_tol": 15, "bursting": false, "chunk_duration": 0, "color_noise_floor": 1, '
@@ -341,7 +343,8 @@ class TestOpenEphysBinaryRecordingInterfaceClassMethodsAndAssertions(RecordingEx
         with self.assertRaisesWith(
             exc_type=ValueError,
             exc_msg=(
-                "Unable to identify the OpenEphys folder structure! Please check that your `folder_path` contains sub-folders of the "
+                "Unable to identify the OpenEphys folder structure! "
+                "Please check that your `folder_path` contains sub-folders of the "
                 "following form: 'experiment<index>' -> 'recording<index>' -> 'continuous'."
             ),
         ):
@@ -354,8 +357,10 @@ class TestOpenEphysBinaryRecordingInterfaceClassMethodsAndAssertions(RecordingEx
         with self.assertRaisesWith(
             exc_type=ValueError,
             exc_msg=(
-                "More than one stream is detected! Please specify which stream you wish to load with the `stream_name` argument. "
-                "To see what streams are available, call `OpenEphysRecordingInterface.get_stream_names(folder_path=...)`."
+                "More than one stream is detected! "
+                "Please specify which stream you wish to load with the `stream_name` argument. "
+                "To see what streams are available, call "
+                " `OpenEphysRecordingInterface.get_stream_names(folder_path=...)`."
             ),
         ):
             OpenEphysBinaryRecordingInterface(
@@ -409,6 +414,23 @@ class TestOpenEphysBinaryRecordingInterfaceVersion0_5_3_Stream2(RecordingExtract
         assert metadata["NWBFile"]["session_start_time"] == datetime(2020, 11, 24, 15, 46, 56)
 
 
+class TestOpenEphysBinaryRecordingInterfaceWithBlocks_version_0_6_block_1_stream_1(
+    RecordingExtractorInterfaceTestMixin, TestCase
+):
+    """From Issue #695, exposed `block_index` argument and added tests on data that include multiple blocks."""
+
+    data_interface_cls = OpenEphysBinaryRecordingInterface
+    interface_kwargs = dict(
+        folder_path=str(DATA_PATH / "openephysbinary" / "v0.6.x_neuropixels_multiexp_multistream" / "Record Node 101"),
+        stream_name="Record Node 101#NI-DAQmx-103.PXIe-6341",
+        block_index=1,
+    )
+    save_directory = OUTPUT_PATH
+
+    def check_extracted_metadata(self, metadata: dict):
+        assert metadata["NWBFile"]["session_start_time"] == datetime(2022, 5, 3, 10, 52, 24)
+
+
 class TestOpenEphysLegacyRecordingInterface(RecordingExtractorInterfaceTestMixin, TestCase):
     data_interface_cls = OpenEphysLegacyRecordingInterface
     interface_kwargs = dict(folder_path=str(DATA_PATH / "openephys" / "OpenEphys_SampleData_1"))
@@ -438,12 +460,12 @@ class TestOpenEphysRecordingInterfaceRouter(RecordingExtractorInterfaceTestMixin
 class TestSpikeGadgetsRecordingInterface(RecordingExtractorInterfaceTestMixin, TestCase):
     data_interface_cls = SpikeGadgetsRecordingInterface
     interface_kwargs = [
-        dict(file_path=str(DATA_PATH / "spikegadgets" / f"20210225_em8_minirec2_ac.rec")),
-        dict(file_path=str(DATA_PATH / "spikegadgets" / f"20210225_em8_minirec2_ac.rec"), gains=[0.195]),
-        dict(file_path=str(DATA_PATH / "spikegadgets" / f"20210225_em8_minirec2_ac.rec"), gains=[0.385] * 512),
-        dict(file_path=str(DATA_PATH / "spikegadgets" / f"W122_06_09_2019_1_fromSD.rec")),
-        dict(file_path=str(DATA_PATH / "spikegadgets" / f"W122_06_09_2019_1_fromSD.rec"), gains=[0.195]),
-        dict(file_path=str(DATA_PATH / "spikegadgets" / f"W122_06_09_2019_1_fromSD.rec"), gains=[0.385] * 128),
+        dict(file_path=str(DATA_PATH / "spikegadgets" / "20210225_em8_minirec2_ac.rec")),
+        dict(file_path=str(DATA_PATH / "spikegadgets" / "20210225_em8_minirec2_ac.rec"), gains=[0.195]),
+        dict(file_path=str(DATA_PATH / "spikegadgets" / "20210225_em8_minirec2_ac.rec"), gains=[0.385] * 512),
+        dict(file_path=str(DATA_PATH / "spikegadgets" / "W122_06_09_2019_1_fromSD.rec")),
+        dict(file_path=str(DATA_PATH / "spikegadgets" / "W122_06_09_2019_1_fromSD.rec"), gains=[0.195]),
+        dict(file_path=str(DATA_PATH / "spikegadgets" / "W122_06_09_2019_1_fromSD.rec"), gains=[0.385] * 128),
     ]
     save_directory = OUTPUT_PATH
 
@@ -468,54 +490,58 @@ class TestSpikeGLXRecordingInterface(RecordingExtractorInterfaceTestMixin, TestC
             manufacturer="Imec",
         )
 
-    def check_electrode_property_helper(self):
-        """Check that the helper function returns in the way the NWB GUIDE table component expects."""
-        electrode_table_json = self.interface.get_electrode_table_json()
 
-        spikeglx_electrode_table_schema = {
-            "type": "array",
-            "minItems": 0,
-            "items": {"$ref": "#/definitions/SpikeGLXElectrodeColumnEntry"},
-            "definitions": {
-                "SpikeGLXElectrodeColumnEntry": {
-                    "type": "object",
-                    "additionalProperties": False,
-                    "required": [
-                        "channel_name",
-                        "contact_shapes",
-                        "gain_to_uV",
-                        "offset_to_uV",
-                        "group",
-                        "group_name",
-                        "inter_sample_shift",
-                        # "location",
-                        "shank_electrode_number",
-                    ],
-                    "properties": {
-                        "channel_name": {"type": "string"},
-                        "contact_shapes": {"type": "string"},
-                        "gain_to_uV": {"type": "number"},
-                        "offset_to_uV": {"type": "number"},
-                        "group": {"type": "number"},
-                        "group_name": {"type": "string"},
-                        "inter_sample_shift": {"type": "number"},
-                        "location": {"type": "array"},
-                        "shank_electrode_number": {"type": "number"},
-                    },
-                }
-            },
-        }
-        print(f"{[electrode_table_json[0]]=}")
-        jsonschema.validate(instance=[electrode_table_json[0]], schema=spikeglx_electrode_table_schema)
+class TestSpikeGLXRecordingInterfaceLongNHP(RecordingExtractorInterfaceTestMixin, TestCase):
+    data_interface_cls = SpikeGLXRecordingInterface
+    interface_kwargs = dict(
+        file_path=str(
+            DATA_PATH
+            / "spikeglx"
+            / "long_nhp_stubbed"
+            / "snippet_g0"
+            / "snippet_g0_imec0"
+            / "snippet_g0_t0.imec0.ap.bin"
+        )
+    )
+    save_directory = OUTPUT_PATH
 
-    def run_custom_checks(self):
-        self.check_electrode_property_helper()
+    def check_extracted_metadata(self, metadata: dict):
+        assert metadata["NWBFile"]["session_start_time"] == datetime(2024, 1, 3, 11, 51, 51)
+        assert metadata["Ecephys"]["Device"][-1] == dict(
+            name="Neuropixel-Imec",
+            description="{"
+            '"probe_type": "1030", '
+            '"probe_type_description": "NP1.0 NHP", '
+            '"flex_part_number": "NPNH_AFLEX_00", '
+            '"connected_base_station_part_number": "NP2_QBSC_00"'
+            "}",
+            manufacturer="Imec",
+        )
 
 
 class TestTdtRecordingInterface(RecordingExtractorInterfaceTestMixin, TestCase):
     data_interface_cls = TdtRecordingInterface
-    interface_kwargs = dict(folder_path=str(DATA_PATH / "tdt" / "aep_05"))
+    test_gain_value = 0.195  # arbitrary value to test gain
+    interface_kwargs = dict(folder_path=str(DATA_PATH / "tdt" / "aep_05"), gain=test_gain_value)
     save_directory = OUTPUT_PATH
+
+    def run_custom_checks(self):
+        # Check that the gain is applied
+        recording_extractor = self.interface.recording_extractor
+        gains = recording_extractor.get_channel_gains()
+        expected_channel_gains = [self.test_gain_value] * recording_extractor.get_num_channels()
+        assert_array_equal(gains, expected_channel_gains)
+
+    def check_read_nwb(self, nwbfile_path: str):
+        from pynwb import NWBHDF5IO
+
+        expected_conversion_factor = self.test_gain_value * 1e-6
+        with NWBHDF5IO(nwbfile_path, "r") as io:
+            nwbfile = io.read()
+            for _, electrical_series in nwbfile.acquisition.items():
+                assert np.isclose(electrical_series.conversion, expected_conversion_factor)
+
+        return super().check_read_nwb(nwbfile_path=nwbfile_path)
 
 
 class TestPlexonRecordingInterface(RecordingExtractorInterfaceTestMixin, TestCase):
@@ -525,3 +551,6 @@ class TestPlexonRecordingInterface(RecordingExtractorInterfaceTestMixin, TestCas
         file_path=str(DATA_PATH / "plexon" / "File_plexon_3.plx"),
     )
     save_directory = OUTPUT_PATH
+
+    def check_extracted_metadata(self, metadata: dict):
+        assert metadata["NWBFile"]["session_start_time"] == datetime(2010, 2, 22, 20, 0, 57)